pyproject.toml
[project]
name = "voice-bot"
version = "0.1.0"
description = "Voice AI SaaS Platform - Real-time voice agents with RAG capabilities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Pipecat with all required services
    "pipecat-ai[deepgram,openai,cartesia,daily]>=0.0.48",
    # FastAPI backend
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # RAG components
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "faiss-cpu>=1.8.0",
    # Document loaders
    "pypdf>=4.0.0",
    "unstructured>=0.15.0",
    "python-docx>=1.1.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiofiles>=24.1.0",
    "httpx>=0.27.0",
    "onnxruntime>=1.23.2",
    # Alternative cheap stack
    "groq>=0.11.0",
    "cartesia>=1.0.0",
    "deepgram-sdk>=3.0.0",
    "elevenlabs>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
